+++ src/index.ts
@@ -1,5 +1,5 @@
 export * from './helpers/usernotes';
 export * from './types/RawUsernotes';
-export * from './classes/UsernotesData';
+export * from './classes/Usernotes';
 export * from './classes/ToolboxClient';
 export * from './subConfig';
